% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellabeat.R
\docType{package}
\name{bellabeat}
\alias{bellabeat}
\title{bellabeat case study package}
\description{
Contains functions for preparing, processing and analyzing the bellabeat case
study.
}
\author{
Simon Müller \email{mueller.si@outlook.com}
}
