% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_into.R
\name{view_into}
\alias{view_into}
\title{Viewing into data}
\usage{
view_into(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Variable
}
\description{
Preform skim, head, colnames and glimpse
}
\examples{
view_into(daily_activity)

}
